src/requirement.txt
autopep8==1.5.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
PyYAML==5.4.1
random-username==1.0.2
raven==6.10.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
xlrd==2.0.1
xlwt==1.3.0
